ingest/job-orchestrator/pyproject.toml
[project]
name = "job_orchestrator"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
readme = "README.md"
dependencies = [
    "dagster==1.12.21",
    "psycopg2-binary",
    "requests",
    "python-dotenv",
    "pandas",
    "pyyaml",
    "openmeteo-requests",
    "requests-cache",
    "retry-requests",
    "numpy",
    "sodapy",
    "ijson",
    "shapely",
    "dagster",
    "dagster-cloud",
    "dagster-dbt",
    "dbt-core<1.12",
    "dbt-postgres<1.12",
    "dbt-postgres<1.12",
    "dbt-postgres<1.12",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "job_orchestrator"
registry_modules = [
    "job_orchestrator.components.*",
]

[tool.dagster]
module_name = "dbt_orchestrator.definitions"
code_location_name = "dbt_orchestrator"

